Add if table_exists?(:sessions) to the drop

+++ priv/repo/migrations/20240321000000_remove_sessions.exs
@@ -2,6 +2,8 @@ defmodule MehrSchulferien.Repo.Migrations.RemoveSessions do
   use Ecto.Migration
 
   def change do
-    drop table(:sessions)
+    if table_exists?(:sessions) do
+      drop table(:sessions)
+    end
   end
 end
